Refactor introspection tests
